Changed the IntPoint and IntRectanble to float version to allow better tweening.

+++ src/Snake.Framework/Texts/Map/MapTextSystem.cs
@@ -33,7 +33,7 @@ namespace Snake.Framework.Texts.Map
 			}
 		}
 
-		public void Draw(int x, int y, string text, string fontName = null)
+		public void Draw(float x, float y, string text, string fontName = null)
 		{
 			((MapFont)GetFont(fontName)).Process(
 				x,
